+++ blast/hit.h
